Split /earn
